tree-wide: use md5 wrapper from a common place
