+++ src/twisted/cred/_digest.py
@@ -11,19 +11,7 @@ Calculations for HTTP Digest authentication.
 
 from binascii import hexlify
 from hashlib import sha1
-
-def md5(data=b'', **kwargs):
-    """
-    Wrapper around hashlib.md5
-    Attempt call with 'usedforsecurity=False' if we get a ValueError, which happens when
-    OpenSSL FIPS mode is enabled:
-    ValueError: error:060800A3:digital envelope routines:EVP_DigestInit_ex:disabled for fips
-    """
-
-    try:
-        return hashlib.md5(data, **kwargs)
-    except ValueError:
-        return hashlib.md5(data, **kwargs, usedforsecurity=False)
+from twisted.python.compat import md5
 
 
 # The digest math
