+++ src/twisted/cred/test/test_digestauth.py
@@ -8,7 +8,6 @@ L{twisted.cred.credentials}.
 
 
 import base64
-import hashlib
 
 from binascii import hexlify
 from hashlib import sha1
@@ -24,22 +23,9 @@ from twisted.cred.credentials import (
 )
 from twisted.cred.error import LoginFailed
 from twisted.internet.address import IPv4Address
+from twisted.python.compat import md5
 from twisted.python.compat import networkString
 from twisted.trial.unittest import TestCase
-
-
-def md5(data=b'', **kwargs):
-    """
-    Wrapper around hashlib.md5
-    Attempt call with 'usedforsecurity=False' if we get a ValueError, which happens when
-    OpenSSL FIPS mode is enabled:
-    ValueError: error:060800A3:digital envelope routines:EVP_DigestInit_ex:disabled for fips
-    """
-
-    try:
-        return hashlib.md5(data, **kwargs)
-    except ValueError:
-        return hashlib.md5(data, **kwargs, usedforsecurity=False)
 
 
 def b64encode(s):
