+++ src/twisted/internet/test/test_unix.py
@@ -6,8 +6,6 @@ Tests for implementations of L{IReactorUNIX}.
 """
 
 
-import hashlib
-
 from os import close, fstat, stat, unlink, urandom
 from pprint import pformat
 from socket import AF_INET, SOCK_STREAM, SOL_SOCKET, socket
@@ -57,26 +55,12 @@ from twisted.internet.test.test_tcp import (
     StreamTransportTestsMixin,
     WriteSequenceTestsMixin,
 )
-from twisted.python.compat import nativeString
+from twisted.python.compat import nativeString, md5
 from twisted.python.failure import Failure
 from twisted.python.filepath import _coerceToFilesystemEncoding
 from twisted.python.log import addObserver, err, removeObserver
 from twisted.python.reflect import requireModule
 from twisted.python.runtime import platform
-
-
-def md5(data=b'', **kwargs):
-    """
-    Wrapper around hashlib.md5
-    Attempt call with 'usedforsecurity=False' if we get a ValueError, which happens when
-    OpenSSL FIPS mode is enabled:
-    ValueError: error:060800A3:digital envelope routines:EVP_DigestInit_ex:disabled for fips
-    """
-
-    try:
-        return hashlib.md5(data, **kwargs)
-    except ValueError:
-        return hashlib.md5(data, **kwargs, usedforsecurity=False)
 
 
 sendmsg = requireModule("twisted.python.sendmsg")
