+++ src/twisted/mail/_cred.py
@@ -6,7 +6,6 @@ Credential managers for L{twisted.mail}.
 """
 
 
-import hashlib
 import hmac
 
 from zope.interface import implementer
@@ -14,7 +13,7 @@ from zope.interface import implementer
 from twisted.cred import credentials
 from twisted.mail._except import IllegalClientResponse
 from twisted.mail.interfaces import IChallengeResponse, IClientAuthentication
-from twisted.python.compat import nativeString
+from twisted.python.compat import md5, nativeString
 
 
 @implementer(IClientAuthentication)
@@ -26,7 +25,7 @@ class CramMD5ClientAuthenticator:
         return b"CRAM-MD5"
 
     def challengeResponse(self, secret, chal):
-        response = hmac.HMAC(secret, chal, digestmod=hashlib.md5).hexdigest()
+        response = hmac.HMAC(secret, chal, digestmod=md5).hexdigest()
         return self.user + b" " + response.encode("ascii")
