+++ src/twisted/mail/test/test_mail.py
@@ -9,7 +9,6 @@ import email.message
 import email.parser
 import errno
 import glob
-import hashlib
 import io
 import os
 import pickle
@@ -53,23 +52,10 @@ from twisted.names import dns
 from twisted.names.dns import Record_CNAME, Record_MX, RRHeader
 from twisted.names.error import DNSNameError
 from twisted.python import failure, log
+from twisted.python.compat import md5
 from twisted.python.filepath import FilePath
 from twisted.python.runtime import platformType
 from twisted.trial.unittest import TestCase
-
-
-def md5(data=b'', **kwargs):
-    """
-    Wrapper around hashlib.md5
-    Attempt call with 'usedforsecurity=False' if we get a ValueError, which happens when
-    OpenSSL FIPS mode is enabled:
-    ValueError: error:060800A3:digital envelope routines:EVP_DigestInit_ex:disabled for fips
-    """
-
-    try:
-        return hashlib.md5(data, **kwargs)
-    except ValueError:
-        return hashlib.md5(data, **kwargs, usedforsecurity=False)
 
 
 @skipIf(platformType != "posix", "twisted.mail only works on posix")
