+++ src/twisted/mail/test/test_pop3.py
@@ -7,7 +7,6 @@ Test cases for L{twisted.mail.pop3} module.
 
 
 import base64
-import hashlib
 import hmac
 import itertools
 from collections import OrderedDict
@@ -28,21 +27,8 @@ from twisted.internet.testing import LineSendingProtocol
 from twisted.mail import pop3
 from twisted.protocols import loopback
 from twisted.python import failure
+from twisted.python.compat import md5
 from twisted.trial import unittest, util
-
-
-def md5(data=b'', **kwargs):
-    """
-    Wrapper around hashlib.md5
-    Attempt call with 'usedforsecurity=False' if we get a ValueError, which happens when
-    OpenSSL FIPS mode is enabled:
-    ValueError: error:060800A3:digital envelope routines:EVP_DigestInit_ex:disabled for fips
-    """
-
-    try:
-        return hashlib.md5(data, **kwargs)
-    except ValueError:
-        return hashlib.md5(data, **kwargs, usedforsecurity=False)
 
 
 class UtilityTests(unittest.SynchronousTestCase):
