+++ src/twisted/spread/pb.py
@@ -29,7 +29,6 @@ To get started, begin with L{PBClientFactory} and L{PBServerFactory}.
 
 
 import random
-import hashlib
 
 from zope.interface import Interface, implementer
 
@@ -45,7 +44,7 @@ from twisted.persisted import styles
 
 # Twisted Imports
 from twisted.python import failure, log, reflect
-from twisted.python.compat import cmp, comparable
+from twisted.python.compat import cmp, comparable, md5
 from twisted.python.components import registerAdapter
 from twisted.spread import banana
 
@@ -79,20 +78,6 @@ from twisted.spread.jelly import _newInstance, globalSecurity, jelly, unjelly
 MAX_BROKER_REFS = 1024
 
 portno = 8787
-
-
-def md5(data=b'', **kwargs):
-    """
-    Wrapper around hashlib.md5
-    Attempt call with 'usedforsecurity=False' if we get a ValueError, which happens when
-    OpenSSL FIPS mode is enabled:
-    ValueError: error:060800A3:digital envelope routines:EVP_DigestInit_ex:disabled for fips
-    """
-
-    try:
-        return hashlib.md5(data, **kwargs)
-    except ValueError:
-        return hashlib.md5(data, **kwargs, usedforsecurity=False)
 
 
 class ProtocolError(Exception):
